requirements.txt
annotated-types==0.7.0
anthropic==0.47.2
anyio==4.8.0
apsw==3.49.0.0
apswutils==0.0.2
beautifulsoup4==4.13.3
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
dnspython==2.7.0
fastcore==1.7.29
fastlite==0.1.1
google-auth==2.39.0
google-genai==1.11.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
jiter==0.8.2
oauthlib==3.2.2
openai==1.64.0
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-fasthtml==0.12.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.45.3
tiktoken==0.9.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==15.0
vercel-blob
